Modified plotting interface for Jupyter Notebooks
Basically modified the phySyncFirmata plotting interface (common_Utility) to common_Utility_2 such that the Matplotlib funcAnimation commands are written in the working Jupyter Notebook instead of the module to allow wider customisation (using Matplotlib's features) and more support for Jupyter Notebooks. 

This file should be pasted in the phySyncFirmata/phyCom directory so that it can be imported. If using virtual environment, the exact path will be (environment_name)/lib/python3.6/site-packages/phySyncFirmata/phyCom